Organization and start of designing.
