+++ package.json
@@ -11,7 +11,8 @@
     "google-map-react": "^2.1.10",
     "react": "^17.0.0",
     "react-dom": "^17.0.0",
-    "react-scripts": "^4.0.0",
+    "react-scripts": "^5.0.0",
+    "styled-components": "^5.3.3",
     "web-vitals": "^2.1.3"
   },
   "scripts": {
